# api/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Infrastructure API
  version: 0.2.0
  description: HTTP contract for the reusable Go backend foundation.
servers:
  - url: /
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Exchange credentials for an access and refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication succeeded.
          headers:
            Set-Cookie:
              description: HttpOnly access and refresh token cookies. The access cookie expires with the JWT access token.
              schema:
                type: string
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            Pragma:
              $ref: "#/components/headers/Pragma"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/refresh:
    post:
      operationId: refreshAccessToken
      summary: Rotate a refresh token and issue a new token pair
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token rotation succeeded.
          headers:
            Set-Cookie:
              description: Rotated HttpOnly access and refresh token cookies. The access cookie expires with the JWT access token.
              schema:
                type: string
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            Pragma:
              $ref: "#/components/headers/Pragma"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the supplied refresh token
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Logout succeeded.
          headers:
            Set-Cookie:
              description: Expired access and refresh token cookies.
              schema:
                type: string
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            Pragma:
              $ref: "#/components/headers/Pragma"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooleanEnvelope"
  /api/v1/users/me:
    get:
      operationId: getCurrentUser
      summary: Return the authenticated user
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users:
    get:
      operationId: listUsers
      summary: List users with optional search and activation filters
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: query
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated user collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      summary: Create a user
      security:
        - bearerAuth: []
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/users/{userID}/active:
    patch:
      operationId: setUserActive
      summary: Activate or deactivate a user
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveRequest"
      responses:
        "200":
          description: User activation state updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooleanEnvelope"
        "422":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/roles:
    get:
      operationId: listRoles
      summary: List roles
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        "200":
          description: Role collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolesEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/permissions:
    get:
      operationId: listPermissions
      summary: List permissions
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        "200":
          description: Permission collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionsEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{userID}/roles/{roleID}:
    post:
      operationId: assignUserRole
      summary: Assign a role to a user
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/RoleID"
      responses:
        "200":
          description: Role assigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooleanEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/roles/{roleID}/permissions/{permissionID}:
    post:
      operationId: grantRolePermission
      summary: Grant a permission to a role
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - $ref: "#/components/parameters/RoleID"
        - $ref: "#/components/parameters/PermissionID"
      responses:
        "200":
          description: Permission granted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooleanEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  headers:
...
    Cache-Control:
      description: Prevents storage of authentication responses.
      schema:
        type: string
        enum: [no-store]
    Pragma:
      description: Prevents caching by HTTP/1.0 intermediaries.
      schema:
        type: string
        enum: [no-cache]
  parameters:
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoleID:
      name: roleID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PermissionID:
      name: permissionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: The authenticated user lacks permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: The request conflicts with an existing resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: HTTP success status repeated for client compatibility.
        msg:
          type: string
          enum: [success]
        data: {}
    ErrorEnvelope:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          description: Stable application error code retained for frontend compatibility.
        msg:
          type: string
          description: Safe public error message.
        data: {}
...
    PaginationMeta:
      type: object
      required: [page, size, total, pages, has_next, has_prev]
      properties:
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
        pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      additionalProperties: false
      properties:
        refresh_token:
          type: string
          minLength: 1
    TokenPair:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          minLength: 1
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          format: int64
          minimum: 1
    TokenPairEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TokenPair"
    User:
      type: object
      required: [id, email, username, display_name, is_active]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        display_name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/User"
    AuthenticatedUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [permissions]
          properties:
            permissions:
              type: array
              items:
                type: string
    AuthenticatedUserEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AuthenticatedUser"
    UserPageEnvelope:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
        msg:
          type: string
          enum: [success]
        data:
          type: object
          required: [items, meta]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/User"
            meta:
              $ref: "#/components/schemas/PaginationMeta"
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
        display_name:
          type: string
    SetActiveRequest:
      type: object
      additionalProperties: false
      required: [is_active]
      properties:
        is_active:
          type: boolean
    Role:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    Permission:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    RolesEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Role"
    PermissionsEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Permission"
    BooleanEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: boolean
...